Break ConvexHull out into a source/header pair
